Cache character images and re-use team banner
Why
---
It's really slow downloading and grabbing character images from marvel
just to get around CORS for canvas to data URL. Also try to reuse team
banner image if it still exists with same characters.

This change addresses the need by
---------------------------------
Cache downloaded character image data. Name banner images with character
IDs separated by "-" so that they can be reused easily if exists.

+++ app/models/team_banner.rb
@@ -1,7 +1,7 @@
 class TeamBanner
 
   IMAGE_FOLDER_PATH     = Rails.root.join('public', 'images', 'banners').freeze
-  TEAM_ID_REPLACE_TOKEN = 'TEAM_ID'
+  TEAM_ID_REPLACE_TOKEN = 'BANNER_KEY'
   IMAGE_NAME_TEMPLATE   = "team_banner_#{TEAM_ID_REPLACE_TOKEN}.png"
   DATA_URL_PREFIX       = 'data:image/png;base64,'
 
@@ -11,20 +11,25 @@ class TeamBanner
     fail ArgumentError, 'Need data attribute'    if attributes[:data].nil?
     fail ArgumentError, 'Need team_id attribute' if attributes[:team_id].nil?
 
-    image_data = attributes[:data].gsub(DATA_URL_PREFIX, '')
-    image_data = image_data.gsub(' ', '+')
-
+    team = Team.find(attributes[:team_id])
     path = IMAGE_FOLDER_PATH.join(
-      IMAGE_NAME_TEMPLATE.gsub(TEAM_ID_REPLACE_TOKEN, attributes[:team_id].to_s)
+      IMAGE_NAME_TEMPLATE.gsub(
+        TEAM_ID_REPLACE_TOKEN, team.characters.pluck(:id).join('-')
+      )
     )
-    File.open(path, 'wb') { |f| f.write(Base64.decode64(image_data)) }
 
-    self.new(Team.find(attributes[:team_id]), path)
+    unless File.exists?(path)
+      image_data = attributes[:data].gsub(DATA_URL_PREFIX, '')
+      image_data = image_data.gsub(' ', '+')
+      File.open(path, 'wb') { |f| f.write(Base64.decode64(image_data)) }
+    end
+
+    self.new(team, path)
   end
 
   def initialize(team, path)
     @team = team
-    @path = path
+    @path = path.to_s
   end
 
 end
